+++ src/rez/rezconfig.py
@@ -289,6 +289,70 @@ variant_select_mode = "version_priority"
 # foo-5+              | Same as range(foo-5+)
 package_filter = None
 
+# Package order. One or more "orderers" can be listed.
+# This will affect the order of version resolution.
+# This can be used to ensure that specific version have priority over others.
+# Higher versions can still be accessed if required.
+#
+# A common use case is to ease migration from python-2 to python-3:
+#
+#     package_orderers = [
+#         {
+#            "type": "per_family",
+#            "orderers": [
+#                 {
+#                     "packages": ["python"],
+#                     "type": "version_split",
+#                     "first_version": "2.7.16"
+#                 }
+#             ]
+#         }
+#     ]
+#
+# This will ensure that for the "python" package, versions equals or lower than "2.7.16" will have priority.
+# Considering the following versions: "2.7.4", "2.7.16", "3.7.4":
+#
+# example             | result
+# --------------------|----------------------------------------------------
+# rez-env python      | python-2.7.16
+# rez-env python-3    | python-3.7.4
+#
+# Here's another example, using another orderer: "soft_timestamp".
+# This orderer will prefer packages released before a provided timestamp.
+# The following example will prefer package released before 2019-09-09.
+#
+#     package_orderers = [
+#         {
+#             "type": "soft_timestamp",
+#             "timestamp": 1568001600,  # 2019-09-09
+#             "rank": 3
+#         }
+#     ]
+#
+# A timestamp can be generated with python:
+#
+#     $ python -c "import datetime, time; print(int(time.mktime(datetime.date(2019, 9, 9).timetuple())))"
+#     1568001600
+#
+# The rank can be used to allow some versions released after the timestamp to still be considered.
+# When using semantic versionnng, a value of 3 is the most common.
+# This will let version with a different patch number to be accepted.
+#
+# Considering a package "foo" with the following versions:
+# - "1.0.0" was released at 2019-09-07
+# - "2.0.0" was released at 2019-09-08
+# - "2.0.1" was released at 2019-09-10
+# - "2.1.0" was released at 2019-09-11
+# - "3.0.0" was released at 2019-09-12
+#
+# example             | timestamp  | rank | result
+# --------------------|------------|---------------------------------------
+# rez-env foo         | 2019-09-09 | 0    | foo-2.0.0
+# rez-env foo         | 2019-09-09 | 3    | foo-2.0.1
+# rez-env foo         | 2019-09-09 | 2    | foo-2.1.0
+# rez-env foo         | 2019-09-09 | 1    | foo-3.0.0
+package_orderers = None
+
 # If True, unversioned packages are allowed. Solve times are slightly better if
 # this value is False.
 allow_unversioned_packages = True
@@ -465,9 +529,9 @@ package_preprocess_mode = "override"
 # is set to an empty string (possibly due to var expansion), it is removed from
 # the message payload.
 #
-
 context_tracking_host = ''
 
+# See [context_tracking_host](#context_tracking_host)
 context_tracking_amqp = {
     "userid": '',
     "password": '',
@@ -477,6 +541,7 @@ context_tracking_amqp = {
     "message_delivery_mode": 1
 }
 
+# See [context_tracking_host](#context_tracking_host)
 context_tracking_context_fields = [
     "status",
     "timestamp",
@@ -488,6 +553,7 @@ context_tracking_context_fields = [
     "resolved_packages"
 ]
 
+# See [context_tracking_host](#context_tracking_host)
 context_tracking_extra_fields = {}
 
 
@@ -704,6 +770,46 @@ create_executable_script_mode = "single"
 # https://pip.pypa.io/en/stable/reference/pip_install/#options
 pip_extra_args = []
 
+# Substitutions for re.sub when unknown parent paths are encountered in the
+# pip package distribution record: *.dist-info/RECORD
+#
+# Rez reads the distribution record to figure out where pip installed files
+# to, then copies them to their final sub-path in the rez package. Ie, python
+# source files are hard coded to be moved under the "python" sub-folder inside
+# the rez package, which then gets added to PYTHONPATH upon rez-env.
+#
+# When it can't find the file listed in the record AND the path starts
+# with a reference to the parent directory "..", the following remaps are
+# used to:
+# 1. Match a path listed in the record to perform the filepath remapping;
+# 2. re.sub expression from step 1 to make the relative path of where pip
+#    actually installed the file to;
+# 3. re.sub expression from step 1 to make the destination filepath, relative
+#    to the rez variant root.
+#
+# Use these tokens to avoid regular expression and OS-specific path issues:
+# - "{pardir}" or "{p}" for parent directory: os.pardir, i.e. ".." on Linux/Mac
+# - "{sep}" or "{s}" for folder separators: os.sep, i.e. "/" on Linux/Mac
+pip_install_remaps = [
+    # Typical bin copy behaviour
+    # Path in record          | pip installed to    | copy to rez destination
+    # ------------------------|---------------------|--------------------------
+    # ../../bin/*             | bin/*               | bin/*
+    {
+        "record_path": r"^{p}{s}{p}{s}(bin{s}.*)",
+        "pip_install": r"\1",
+        "rez_install": r"\1",
+    },
+    # Fix for https://github.com/nerdvegas/rez/issues/821
+    # Path in record          | pip installed to    | copy to rez destination
+    # ------------------------|---------------------|--------------------------
+    # ../../lib/python/*      | *                   | python/*
+    {
+        "record_path": r"^{p}{s}{p}{s}lib{s}python{s}(.*)",
+        "pip_install": r"\1",
+        "rez_install": r"python{s}\1",
+    },
+]
 
 ###############################################################################
 # Rez-1 Compatibility